+++ templates/parts/settings/people.php
@@ -54,6 +54,8 @@ $isNcAdmin = !empty($_['isNcAdmin']);
 	</ul>
 
 	<?php if ($isNcAdmin): ?>
-		<button type="submit" class="lck-btn lck-btn--primary"><?php p($l->t('Save')); ?></button>
+		<div class="lck-form-actions">
+			<button type="submit" class="lck-btn lck-btn--primary"><?php p($l->t('Save')); ?></button>
+		</div>
 	<?php endif; ?>
 </form>